Create base Unarmed Override Controller
Starting to get ready for combat. To do that each weapon type will need an override controller for the animations. To make this easier I replaced all of the animations with well named animations and created an overridecontroller editor. Base character animators should now be set to the Unarmed override controller.

Note: All animations in the Unarmed override controller should be filled. If not the player will jump to an empty animation state.